+++ libs/base_core/src/text/char/digits/usize.rs
@@ -1,7 +1,7 @@
 // devela_base_core::text::char::digits::usize
 
 use super::*;
-use crate::{Cmp, Lut, StringU8, is};
+use crate::{Cmp, StringU8, TextLut, is};
 
 impl Digits<usize> {
     /// The maximum number of decimal digits a `usize` can represent in the current platform.
@@ -38,7 +38,7 @@ impl Digits<usize> {
     #[inline(always)]
     pub const fn digit_at_index10(self, index: u8) -> u8 {
         is![index >= self.count_digits10(); return b'0'];
-        let power = Lut::POWERS10[index as usize] as usize;
+        let power = TextLut::POWERS10[index as usize] as usize;
         (self.0 / power % 10) as u8 + b'0'
     }
 
@@ -48,7 +48,7 @@ impl Digits<usize> {
     #[must_use]
     pub const fn digit_at_index10_checked(self, index: u8) -> Option<u8> {
         is![index >= self.count_digits10(); return None];
-        let power = Lut::POWERS10[index as usize] as usize;
+        let power = TextLut::POWERS10[index as usize] as usize;
         Some((self.0 / power % 10) as u8 + b'0')
     }
 
@@ -59,7 +59,7 @@ impl Digits<usize> {
     pub const fn digit_at_index16(self, index: u8) -> u8 {
         let shift = index as u32 * 4;
         let digit = self.0.unbounded_shr(shift) & 0xF;
-        Lut::DIGITS_BASE36[digit]
+        TextLut::DIGITS_BASE36[digit]
     }
 
     /// Returns `Some(ASCII digit)` if the index is within the number's hexadecimal digits,
@@ -70,7 +70,7 @@ impl Digits<usize> {
         is![index >= self.count_digits16(); return None];
         let shift = index as u32 * 4;
         let digit = self.0.unbounded_shr(shift) & 0xF;
-        Some(Lut::DIGITS_BASE36[digit])
+        Some(TextLut::DIGITS_BASE36[digit])
     }
 
     /* digit_value_at_ */
@@ -81,7 +81,7 @@ impl Digits<usize> {
     #[must_use]
     pub const fn digit_value_at_index10(self, index: u8) -> u8 {
         is![index >= self.count_digits10(); return 0];
-        let power = Lut::POWERS10[index as usize] as usize;
+        let power = TextLut::POWERS10[index as usize] as usize;
         (self.0 / power % 10) as u8
     }
     /// Returns `Some(numeric_value)` (0-9) of the decimal digit at the specified index.
@@ -90,7 +90,7 @@ impl Digits<usize> {
     #[must_use]
     pub const fn digit_value_at_index10_checked(self, index: u8) -> Option<u8> {
         is![index >= self.count_digits10(); return None];
-        let power = Lut::POWERS10[index as usize] as usize;
+        let power = TextLut::POWERS10[index as usize] as usize;
         Some((self.0 / power % 10) as u8)
     }