+++ libs/base_core/src/text/char/namespace/u32.rs
@@ -1,6 +1,6 @@
 // devela_base_core::text::char::namespace::u32
 
-use crate::{Char, Lut};
+use crate::{Char, TextLut};
 
 /// # Methods over `u32`.
 #[rustfmt::skip]
@@ -318,7 +318,7 @@ impl Char<u32> {
     /// Returns the ASCII `&'static str` representation of the value, or `""` if non-ASCII.
     #[must_use]
     pub const fn as_ascii(self) -> &'static str {
-        if self.is_ascii() { Lut::ASCII_CHARS[self.0 as usize] } else { "" }
+        if self.is_ascii() { TextLut::ASCII_CHARS[self.0 as usize] } else { "" }
     }
 
     /// Returns the ASCII `&'static str` representation of the value, or panics if non-ASCII.
@@ -326,7 +326,7 @@ impl Char<u32> {
     /// # Panics
     /// Panics if the character is not ASCII.
     #[must_use]
-    pub const fn as_ascii_unchecked(self) -> &'static str { Lut::ASCII_CHARS[self.0 as usize] }
+    pub const fn as_ascii_unchecked(self) -> &'static str { TextLut::ASCII_CHARS[self.0 as usize] }
 
     /// Converts the Unicode scalar value to a UTF-8 encoded byte sequence array.
     ///