+++ libs/base_core/src/text/char/scalar/c7.rs
@@ -4,7 +4,7 @@
 //
 
 use super::*;
-use crate::{Char, CharAscii, DataOverflow, Lut, NonExtremeU8};
+use crate::{Char, CharAscii, DataOverflow, NonExtremeU8, TextLut};
 
 impl char7 {
     /* private helper fns */
@@ -104,7 +104,7 @@ impl char7 {
     /// Returns the string slice representation.
     #[inline(always)]
     pub const fn to_str(&self) -> &'static str {
-        Lut::ASCII_CHARS[self.to_byte() as usize]
+        TextLut::ASCII_CHARS[self.to_byte() as usize]
     }
 
     /// Converts a `char7` to `CharAscii`.