+++ libs/base_core/src/text/lut.rs
@@ -1,0 +1,11 @@
+// devela_base_core::text::lut
+//
+//! Defines the [`TextLut`] namespace.
+//
+
+#[doc = crate::_TAG_CODE!()]
+#[doc = crate::_TAG_NAMESPACE!()]
+#[doc = concat!["A text-related ", crate::_ABBR_LUT!() , " namespace."]]
+#[doc = crate::_doc_location!("text")]
+#[derive(Debug)]
+pub struct TextLut;